rename logger to terminal
