+++ src/Terminal.ts
