+++ src/index.ts
@@ -3,7 +3,7 @@ export type { Command } from './Command.type';
 export type { Configuration } from './Configuration.type';
 
 export { read_boolean_option, read_numerical_option, read_string_option} from './Argv';
-export { log, Logger } from './Logger';
+export { log, Logger } from './Terminal';
 export { print_help } from './print_help';
 export { print_version } from './print_version';
 export { run_command_with_options } from './run_command_with_options';
