+++ src/print_did_you_mean.ts
@@ -1,6 +1,6 @@
 import { nice_executable_name } from './Argv';
 import type { Command } from './Command.type';
-import { log } from './Logger';
+import { log } from './Terminal';
 
 import { nearest_string } from './nearest_string';
