+++ src/print_version.ts
@@ -1,5 +1,5 @@
 import type { Configuration } from './Configuration.type';
-import { log } from './Logger';
+import { log } from './Terminal';
 import * as style from './style';
 
 export function print_version(executable: string, cfg: Configuration): void {
